refactor: 아바타 기능 재화소비코드 리펙토링
avatar 의상구매 시 재화소비코드를 coin package에서 가져옴
coinMapper의 substarct 코드를 amount만 빼도록 수정

closes #126

+++ src/main/java/org/scoula/avatar/service/AvatarService.java
@@ -17,5 +17,5 @@ public interface AvatarService{
 
     void insertClothe(Long userId, Long itemId);
 
-    Long getCoin(Long userId);
+    int getCoin(Long userId);
 }
